requirements.txt
asgiref==3.2.10
Babel==2.8.0
cycler==0.10.0
Django==3.0.8
django-bootstrap-breadcrumbs==0.9.2
django-crispy-forms==1.9.1
django-phonenumber-field==4.0.0
gunicorn==20.0.4
kiwisolver==1.2.0
lxml==4.5.1
matplotlib==3.2.2
numpy==1.19.0
phonenumbers==8.12.6
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
